警告を修正

+++ Object/Collision.cpp
@@ -68,7 +68,7 @@ void Collision::Update()
 
 	//炎玉エネミー判定
 	m_pEnemy->GetHitFireBallFall(HitEnemyFireBall()); //重力.y
-	m_pEnemy->GetFireBallPos(static_cast<float>(m_enemyFireBallPosY)); //位置.y
+	m_pEnemy->GetFireBallPos(static_cast<int>(m_enemyFireBallPosY)); //位置.y
 
 	//落ちるエネミー判定
 	m_pEnemy->GetFallenRange(HitFallenRange());//範囲に入っているかどうか
@@ -611,7 +611,7 @@ int Collision::HitEnemyFireBall()
 		if ((Stage1::kBoxBottom1Y > m_pEnemy->GetTop()) &&
 			(Stage1::kBox1Y < m_pEnemy->GetBottom()))
 		{
-			m_enemyFireBallPosY = Stage1::kBox1Y - kEnemyPosY + kPlayerPosPulsY;
+			m_enemyFireBallPosY = Stage1::kBox1Y - static_cast<int>(kEnemyPosY + kPlayerPosPulsY);
 			return 2;
 		}
 	}
@@ -622,7 +622,7 @@ int Collision::HitEnemyFireBall()
 		if ((Stage1::kBoxBottom2Y > m_pEnemy->GetTop()) &&
 			(Stage1::kBox2Y < m_pEnemy->GetBottom()))
 		{
-			m_enemyFireBallPosY = Stage1::kBox2Y - kEnemyPosY + kPlayerPosPulsY;
+			m_enemyFireBallPosY = Stage1::kBox2Y - static_cast<int>(kEnemyPosY + kPlayerPosPulsY);
 			return 2;
 		}
 	}
@@ -633,7 +633,7 @@ int Collision::HitEnemyFireBall()
 		if ((Stage1::kBoxBottom3Y > m_pEnemy->GetTop()) &&
 			(Stage1::kBox3Y < m_pEnemy->GetBottom()))
 		{
-			m_enemyFireBallPosY = Stage1::kBox3Y - kEnemyPosY + kPlayerPosPulsY;
+			m_enemyFireBallPosY = Stage1::kBox3Y - static_cast<int>(kEnemyPosY + kPlayerPosPulsY);
 			return 2;
 		}
 	}
@@ -644,7 +644,7 @@ int Collision::HitEnemyFireBall()
 		if ((Stage1::kBoxBottom4Y > m_pEnemy->GetTop()) &&
 			(Stage1::kBox4Y < m_pEnemy->GetBottom()))
 		{
-			m_enemyFireBallPosY = Stage1::kBox4Y - kEnemyPosY + kPlayerPosPulsY;
+			m_enemyFireBallPosY = Stage1::kBox4Y - static_cast<int>(kEnemyPosY + kPlayerPosPulsY);
 			return 2;
 		}
 	}
@@ -655,7 +655,7 @@ int Collision::HitEnemyFireBall()
 		if ((Stage1::kBoxBottom5Y > m_pEnemy->GetTop()) &&
 			(Stage1::kBox5Y < m_pEnemy->GetBottom()))
 		{
-			m_enemyFireBallPosY = Stage1::kBox5Y - kEnemyPosY + kPlayerPosPulsY;
+			m_enemyFireBallPosY = Stage1::kBox5Y - static_cast<int>(kEnemyPosY + kPlayerPosPulsY);
 			return 2;
 		}
 	}
@@ -666,7 +666,7 @@ int Collision::HitEnemyFireBall()
 		if ((Stage1::kBoxBottom6Y > m_pEnemy->GetTop()) &&
 			(Stage1::kBox6Y < m_pEnemy->GetBottom()))
 		{
-			m_enemyFireBallPosY = Stage1::kBox6Y - kEnemyPosY + kPlayerPosPulsY;
+			m_enemyFireBallPosY = Stage1::kBox6Y - static_cast<int>(kEnemyPosY + kPlayerPosPulsY);
 			return 2;
 		}
 	}
@@ -677,7 +677,7 @@ int Collision::HitEnemyFireBall()
 		if ((Stage1::kBoxBottom7Y > m_pEnemy->GetTop()) &&
 			(Stage1::kBox7Y < m_pEnemy->GetBottom()))
 		{
-			m_enemyFireBallPosY = Stage1::kBox7Y - kEnemyPosY + kPlayerPosPulsY;
+			m_enemyFireBallPosY = Stage1::kBox7Y - static_cast<int>(kEnemyPosY + kPlayerPosPulsY);
 			return 2;
 		}
 	}
@@ -688,7 +688,7 @@ int Collision::HitEnemyFireBall()
 		if ((Stage1::kBoxBottom8Y > m_pEnemy->GetTop()) &&
 			(Stage1::kBox8Y < m_pEnemy->GetBottom()))
 		{
-			m_enemyFireBallPosY = Stage1::kBox8Y - kEnemyPosY + kPlayerPosPulsY;
+			m_enemyFireBallPosY = Stage1::kBox8Y - static_cast<int>(kEnemyPosY + kPlayerPosPulsY);
 			return 2;
 		}
 	}
