+++ Object/Collision3.cpp
@@ -79,7 +79,7 @@ void Collision3::Update()
 
 	//炎玉エネミー判定
 	m_pEnemy->GetHitFireBallFall(HitEnemyFireBall()); //重力.y
-	m_pEnemy->GetFireBallPos(static_cast<float>(m_enemyFireBallPosY)); //位置.y
+	m_pEnemy->GetFireBallPos(static_cast<int>(m_enemyFireBallPosY)); //位置.y
 
 	//落ちるエネミー判定
 	m_pEnemy->GetFallenRange(HitFallenRange());//範囲に入っているかどうか
@@ -623,7 +623,7 @@ int Collision3::HitEnemyFireBall()
 		if ((Stage2::kBoxBottom1Y > m_pEnemy->GetTop()) &&
 			(Stage2::kBox1Y < m_pEnemy->GetBottom()))
 		{
-			m_enemyFireBallPosY = Stage2::kBox1Y - kEnemyPosY + kPlayerPosPulsY;
+			m_enemyFireBallPosY = Stage2::kBox1Y - static_cast<int>(kEnemyPosY + kPlayerPosPulsY);
 			return 2;
 		}
 	}
@@ -634,7 +634,7 @@ int Collision3::HitEnemyFireBall()
 		if ((Stage2::kBoxBottom2Y > m_pEnemy->GetTop()) &&
 			(Stage2::kBox2Y < m_pEnemy->GetBottom()))
 		{
-			m_enemyFireBallPosY = Stage2::kBox2Y - kEnemyPosY + kPlayerPosPulsY;
+			m_enemyFireBallPosY = Stage2::kBox2Y - static_cast<int>(kEnemyPosY + kPlayerPosPulsY);
 			return 2;
 		}
 	}
@@ -645,7 +645,7 @@ int Collision3::HitEnemyFireBall()
 		if ((Stage2::kBoxBottom3Y > m_pEnemy->GetTop()) &&
 			(Stage2::kBox3Y < m_pEnemy->GetBottom()))
 		{
-			m_enemyFireBallPosY = Stage2::kBox3Y - kEnemyPosY + kPlayerPosPulsY;
+			m_enemyFireBallPosY = Stage2::kBox3Y - static_cast<int>(kEnemyPosY + kPlayerPosPulsY);
 			return 2;
 		}
 	}
@@ -656,7 +656,7 @@ int Collision3::HitEnemyFireBall()
 		if ((Stage2::kBoxBottom4Y > m_pEnemy->GetTop()) &&
 			(Stage2::kBox4Y < m_pEnemy->GetBottom()))
 		{
-			m_enemyFireBallPosY = Stage2::kBox4Y - kEnemyPosY + kPlayerPosPulsY;
+			m_enemyFireBallPosY = Stage2::kBox4Y - static_cast<int>(kEnemyPosY + kPlayerPosPulsY);
 			return 2;
 		}
 	}
@@ -667,7 +667,7 @@ int Collision3::HitEnemyFireBall()
 		if ((Stage2::kBoxBottom5Y > m_pEnemy->GetTop()) &&
 			(Stage2::kBox5Y < m_pEnemy->GetBottom()))
 		{
-			m_enemyFireBallPosY = Stage2::kBox5Y - kEnemyPosY + kPlayerPosPulsY;
+			m_enemyFireBallPosY = Stage2::kBox5Y - static_cast<int>(kEnemyPosY + kPlayerPosPulsY);
 			return 2;
 		}
 	}
@@ -678,7 +678,7 @@ int Collision3::HitEnemyFireBall()
 		if ((Stage2::kBoxBottom6Y > m_pEnemy->GetTop()) &&
 			(Stage2::kBox6Y < m_pEnemy->GetBottom()))
 		{
-			m_enemyFireBallPosY = Stage2::kBox6Y - kEnemyPosY + kPlayerPosPulsY;
+			m_enemyFireBallPosY = Stage2::kBox6Y - static_cast<int>(kEnemyPosY + kPlayerPosPulsY);
 			return 2;
 		}
 	}
@@ -689,7 +689,7 @@ int Collision3::HitEnemyFireBall()
 		if ((Stage2::kBoxBottom7Y > m_pEnemy->GetTop()) &&
 			(Stage2::kBox7Y < m_pEnemy->GetBottom()))
 		{
-			m_enemyFireBallPosY = Stage2::kBox7Y - kEnemyPosY + kPlayerPosPulsY;
+			m_enemyFireBallPosY = Stage2::kBox7Y - static_cast<int>(kEnemyPosY + kPlayerPosPulsY);
 			return 2;
 		}
 	}
@@ -700,7 +700,7 @@ int Collision3::HitEnemyFireBall()
 		if ((Stage2::kBoxBottom8Y > m_pEnemy->GetTop()) &&
 			(Stage2::kBox8Y < m_pEnemy->GetBottom()))
 		{
-			m_enemyFireBallPosY = Stage2::kBox8Y - kEnemyPosY + kPlayerPosPulsY;
+			m_enemyFireBallPosY = Stage2::kBox8Y - static_cast<int>(kEnemyPosY + kPlayerPosPulsY);
 			return 2;
 		}
 	}
