+++ Object/TitleCursorGame.cpp
@@ -7,8 +7,8 @@ TitleCursorGame::TitleCursorGame() :
 	m_mouseY(0),
 	m_padPos(0.0f, 0.0f)
 {
-	m_padPos.x = Game::kScreenWidth / 2;
-	m_padPos.y = Game::kScreenHeight / 2 + 30;
+	m_padPos.x = static_cast<float>(Game::kScreenWidth) / 2;
+	m_padPos.y = static_cast<float>(Game::kScreenHeight / 2 + 30);
 }
 
 TitleCursorGame::~TitleCursorGame()
@@ -58,18 +58,18 @@ void TitleCursorGame::Draw()
 //プレイヤーの行動範囲
 void TitleCursorGame::PosSet()
 {
-	if (m_padPos.x < 0.0f) m_padPos.x = 1.0f;
-	if (m_padPos.x > Game::kScreenWidth) m_padPos.x = static_cast<float>(Game::kScreenWidth) - 1.0f;
-	if (m_padPos.y < 0.0f) m_padPos.y = 1.0f;
-	if (m_padPos.y > Game::kScreenHeight) m_padPos.y = static_cast<float>(Game::kScreenHeight) - 1.0f;
+	if (static_cast<int>(m_padPos.x) < 0.0f) static_cast<int>(m_padPos.x = 1.0f);
+	if (static_cast<int>(m_padPos.x) > Game::kScreenWidth) m_padPos.x = static_cast<float>(Game::kScreenWidth) - 1.0f;
+	if (static_cast<int>(m_padPos.y) < 0.0f) static_cast<int>(m_padPos.y = 1.0f);
+	if (static_cast<int>(m_padPos.y) > Game::kScreenHeight) m_padPos.y = static_cast<float>(Game::kScreenHeight) - 1.0f;
 }
 
 void TitleCursorGame::DrawPadCursor()
 {
-	DrawCircle(m_padPos.x, m_padPos.y, 20, 0x00ffff, false);
+	DrawCircle(static_cast<int>(m_padPos.x), static_cast<int>(m_padPos.y), 20, 0x00ffff, false);
 	// 描画ブレンドモードをアルファブレンド（５０％）にする
 	SetDrawBlendMode(DX_BLENDMODE_ALPHA, 128);
-	DrawCircle(m_padPos.x, m_padPos.y, 19, 0xffff00, true);
+	DrawCircle(static_cast<int>(m_padPos.x), static_cast<int>(m_padPos.y), 19, 0xffff00, true);
 	// 描画ブレンドモードをノーブレンドにする
 	SetDrawBlendMode(DX_BLENDMODE_NOBLEND, 0);
 }
