+++ Scene/SceneMain.cpp
@@ -157,12 +157,12 @@ void SceneMain::Draw()
 		m_pEnemy->Draw();
 	}
 
-	SetDrawBlendMode(DX_BLENDMODE_ALPHA, m_fadeValue);
+	SetDrawBlendMode(DX_BLENDMODE_ALPHA, static_cast<int>(m_fadeValue));
 	DrawBox(0, 0, 640, 480, 0xffffff, true);
 	SetDrawBlendMode(DX_BLENDMODE_NOBLEND, 0);
 
 	//フェイド処理
-	SetDrawBlendMode(DX_BLENDMODE_ALPHA, m_fadeValue);
+	SetDrawBlendMode(DX_BLENDMODE_ALPHA, static_cast<int>(m_fadeValue));
 	DrawBox(0, 0, Game::kScreenWidth, Game::kScreenHeight, 0x000000, true);
 	SetDrawBlendMode(DX_BLENDMODE_NOBLEND, 0);
 }
